fix: an async generator code gen issue
If an assign node's expr contains any awaits, the await code gen shall hapen before the left symbol resolution, or the symbol may not dominant all uses.
This fix simply fix this issue by resolve symbol twice, this is, of cource, ugly, but it's the simplest way I can think of.